requirements.txt
openai
PyQt5
pandas
pdfplumber
pydantic
pyinstaller
python-dotenv
openpyxl
pdf2image
spacy
pytesseract 
pillow
PyPDF2
PyMuPDF
python-docx
docx
